Update for November 2020 rules
This updates the nodesByUserAgentInConsensus function for the
November 2020 upgrade. This should be close to the coin.dance
numbers.

This removes the bchd version check (which was wrong, 'bchd' should
be lowercase) because there is no released version with support
for ASERT yet.

This also removes counting for unknown nodeTypes. It is wrong to
count clients like Bitcoin XT to be in consensus.